+++ tests/unit/neptune/new/internal/backends/test_hosted_neptune_backend.py
@@ -352,6 +352,8 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                             parent_identifier=str(container_id),
                             entries=[("/path/to/file", "/path/to")],
                             default_request_params=DEFAULT_REQUEST_KWARGS,
+                            exclude_directory_files=True,
+                            exclude_metadata_from_hash=True,
                         ),
                         call(
                             swagger_client=swagger_client_wrapper,
@@ -363,6 +365,8 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                                 ("/path/to/file2", None),
                             ],
                             default_request_params=DEFAULT_REQUEST_KWARGS,
+                            exclude_directory_files=True,
+                            exclude_metadata_from_hash=True,
                         ),
                         call(
                             swagger_client=swagger_client_wrapper,
@@ -374,6 +378,8 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                                 ("/path/to/file2", None),
                             ],
                             default_request_params=DEFAULT_REQUEST_KWARGS,
+                            exclude_directory_files=True,
+                            exclude_metadata_from_hash=True,
                         ),
                     ],
                     any_order=True,
@@ -445,6 +451,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                             parent_identifier=str(container_id),
                             entries=[("/path/to/file", "/path/to")],
                             default_request_params=DEFAULT_REQUEST_KWARGS,
+                            exclude_directory_files=True,
                         ),
                         call(
                             swagger_client=swagger_client_wrapper,
@@ -457,6 +464,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                                 ("/path/to/file2", None),
                             ],
                             default_request_params=DEFAULT_REQUEST_KWARGS,
+                            exclude_directory_files=True,
                         ),
                         call(
                             swagger_client=swagger_client_wrapper,
@@ -469,6 +477,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                                 ("/path/to/file2", None),
                             ],
                             default_request_params=DEFAULT_REQUEST_KWARGS,
+                            exclude_directory_files=True,
                         ),
                     ],
                     any_order=True,